+++ crates/feagi-brain-development/src/connectivity/core_morphologies/patterns.rs
@@ -21,8 +21,8 @@ pub fn apply_patterns_morphology(
     src_area_id: u32,
     dst_area_id: u32,
     patterns: Vec<(Pattern3D, Pattern3D)>,
-    weight: u8,
-    psp: u8,
+    weight: f32,
+    psp: f32,
     synapse_attractivity: u8,
     synapse_type: SynapseType,
 ) -> BduResult<u32> {
