+++ crates/feagi-brain-development/src/connectivity/core_morphologies/rotator_z.rs
@@ -21,8 +21,8 @@ pub fn apply_rotator_z_morphology_with_dimensions(
     dst_area_id: u32,
     src_dimensions: (usize, usize, usize),
     dst_dimensions: (usize, usize, usize),
-    weight: u8,
-    psp: u8,
+    weight: f32,
+    psp: f32,
     synapse_attractivity: u8,
     synapse_type: SynapseType,
 ) -> BduResult<u32> {
